server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
psycopg2-binary
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
stripe==6.7.0
reportlab==4.0.4
openpyxl==3.1.2
Pillow==10.0.1
requests==2.31.0
scrapy==2.11.0
scrapy-redis==0.7.3
pymongo==4.5.0
fake-useragent==1.4.0
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-redis==6.0.0